Add files via upload
Sometimes, it would seem, my shower thoughts are of important issues, and how one mught explore them. In this case, I am aware of the (redundant) discourse surrounding the existence of the 'Gender Pay Gap' (GPG). Let me be clear, that is not to say it is redundant to explore this issue and the ways in which it may be addressed, but rather, it is redundant to 'debate' its existence. It was the very fact that people will argue over it's existence that inspired me to explore the empirical proof behind the GPG, such that I can contribute to the ending of the 'debates' to expedite moving onto the part where solutions are found! For, in the  words of one such GPG-denying pundits of the manosphere: "Facts don't care about your feelings".

This is NOT an extensive exploration of the existence of the GPG, nor will it explain the reasons behind it, rather  it is an exercise in data science using a dataset I find important and interesting.